refactor vmss instance interfaces
